# docs/swagger.yaml
basePath: /
definitions:
  github_com_proarash_ecommerce-server_internal_finance.Order:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      discount_amount:
        type: number
      discount_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_finance.OrderItem'
        type: array
      status:
        enum:
        - pending
        - paid
        - failed
        - cancelled
        - processing
        - delivered
        type: string
      total_amount:
        type: number
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  github_com_proarash_ecommerce-server_internal_finance.OrderItem:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      order_id:
        type: integer
      product:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      unit_price:
        type: number
      updatedAt:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_media.Media:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      media_type:
        type: string
      mime_type:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Attribute:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Category:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      media:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_media.Media'
      media_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_product.Product:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Attribute'
        type: array
      category:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Category'
      category_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      media:
        items:
          $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_media.Media'
        type: array
      price:
        type: number
      sku:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_proarash_ecommerce-server_internal_types.ApiResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  github_com_proarash_ecommerce-server_internal_types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_cart.AddItemDto:
    properties:
      product_id:
        minimum: 1
        type: integer
      quantity:
        maximum: 1000
        minimum: 1
        type: integer
    required:
    - product_id
    type: object
  internal_cart.CartItem:
    properties:
      cart_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      product:
        $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_product.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  internal_cart.CartResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_cart.CartItem'
        type: array
      total_amount:
        type: number
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  internal_cart.CheckoutDto:
    properties:
      discount_code:
        example: SUMMER30
        maxLength: 32
        type: string
    type: object
  internal_cart.UpdateItemDto:
    properties:
      quantity:
        maximum: 1000
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: 'E-commerce backend: catalog, cart, orders, Zibal payments, inventory,
    CMS, support chat and notifications. All JSON responses are wrapped in {"message":
    string, "data": any}.'
  title: E-Commerce Server API
  version: "1.0"
paths:
...
  /cart:
    get:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get current cart
      tags:
      - Cart
  /cart/checkout:
    post:
      consumes:
      - application/json
      description: Converts the cart into a pending order, optionally redeeming a
        discount code, issues a system pre-invoice and sends invoice/pre-invoice bot
        messages
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Optional discount code
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_cart.CheckoutDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_finance.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Checkout cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Adds the product or increases its quantity if already present
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_cart.AddItemDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Add product to cart
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      parameters:
      - description: Session to authenticate with; selects the staff or customer auth
          cookie
        enum:
        - staff
        - customer
        in: query
        name: user_type
        type: string
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_cart.UpdateItemDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_cart.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_proarash_ecommerce-server_internal_types.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"